vision_api/requirements.txt
torch>=2.0.0
open_clip_torch>=2.20.0
pydicom>=2.4.0
pillow>=10.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
numpy>=1.24.0
python-multipart>=0.0.6 